Bring back example application
And capture _all_ plugin providers in it.
