+++ .env.example
@@ -1,0 +1,36 @@
+#
+# Copy this file to .env and update the credentials for the providers you are
+# trying to test. These variables must all have non-empty values for the
+# application to boot, but you only need to set real values for those Providers
+# you plan to try.
+#
+###
+BATTLE_NET_CLIENT_ID=x
+BATTLE_NET_CLIENT_SECRET=x
+
+BITBUCKET_CLIENT_ID=x
+BITBUCKET_CLIENT_SECRET=x
+
+EVE_ONLINE_CLIENT_ID=x
+EVE_ONLINE_CLIENT_SECRET=x
+
+GITHUB_CLIENT_ID=x
+GITHUB_CLIENT_SECRET=x
+
+GOOGLE_CLIENT_ID=x
+GOOGLE_CLIENT_SECRET=x
+
+NYLAS_CLIENT_ID=x
+NYLAS_CLIENT_SECRET=x
+
+SALES_FORCE_CLIENT_ID=x
+SALES_FORCE_CLIENT_SECRET=x
+
+SLACK_CLIENT_ID=x
+SLACK_CLIENT_SECRET=x
+
+SPOTIFY_CLIENT_ID=x
+SPOTIFY_CLIENT_SECRET=x
+
+UPCASE_CLIENT_ID=x
+UPCASE_CLIENT_SECRET=x
