+++ package.yaml
@@ -34,6 +34,27 @@ library:
     - yesod-auth >=1.3 && <1.5
     - yesod-core >=1.2 && <1.5
 
+executables:
+  yesod-auth-oauth2-example:
+    main: Main.hs
+    source-dirs: example
+    ghc-options:
+      - -threaded
+      - -rtsopts
+      - -with-rtsopts=-N
+    dependencies:
+      - yesod-auth-oauth2
+      - containers
+      - http-conduit
+      - load-env
+      - text
+      - warp
+      - yesod
+      - yesod-auth
+    when:
+      - condition: ! '!(flag(example))'
+        buildable: false
+
 tests:
   test:
     main: Spec.hs
@@ -42,3 +63,9 @@ tests:
       - yesod-auth-oauth2
       - hspec
       - uri-bytestring
+
+flags:
+  example:
+    description: Build the example application
+    manual: false
+    default: false
